mission: 3주차 리팩토링
