+++ mission/week3/app/src/main/AndroidManifest.xml
@@ -13,10 +13,10 @@
         android:theme="@style/Theme.Flo"
         tools:targetApi="31">
         <activity
-            android:name=".SongActivity"
+            android:name=".ui.player.SongActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
