+++ mission/week3/app/src/main/java/com/example/week3/model/Song.kt
@@ -1,6 +1,4 @@
-package com.example.week3
-
-import androidx.annotation.DrawableRes
+package com.example.week3.model
 
 data class Song(
     val title: String = "",
