+++ mission/week3/app/src/main/java/com/example/week3/ui/album/AlbumFragment.kt
@@ -1,15 +1,15 @@
-package com.example.week3.fragment
+package com.example.week3.ui.album
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.viewpager2.widget.ViewPager2
-import com.example.week3.adapter.AlbumVPAdapter
-import com.example.week3.MainActivity
+import com.example.week3.ui.MainActivity
 import com.example.week3.R
 import com.example.week3.databinding.FragmentAlbumBinding
+import com.example.week3.ui.main.HomeFragment
 import com.google.android.material.tabs.TabLayoutMediator
 
 class AlbumFragment : Fragment() {
@@ -24,6 +24,22 @@ class AlbumFragment : Fragment() {
     ): View {
         binding = FragmentAlbumBinding.inflate(inflater, container, false)
 
+        setupBackBtnClickListener()
+
+        val albumAdapter = AlbumVPAdapter(this)
+        binding.albumContentVp.apply {
+            adapter = albumAdapter
+            orientation = ViewPager2.ORIENTATION_HORIZONTAL
+        }
+
+        TabLayoutMediator(binding.albumContentTb, binding.albumContentVp) { tab, position ->
+            tab.text = tabInfo[position] // 탭뷰의 아이템 이름 설정
+        }.attach() // 탭뷰와 vp 연결
+
+        return binding.root
+    }
+
+    private fun setupBackBtnClickListener() {
         binding.albumBackIv.setOnClickListener {
             (context as MainActivity)
                 .supportFragmentManager
@@ -31,15 +47,5 @@ class AlbumFragment : Fragment() {
                 .replace(R.id.main_frm, HomeFragment())
                 .commitAllowingStateLoss()
         }
-
-        val albumAdapter = AlbumVPAdapter(this)
-        binding.albumContentVp.adapter = albumAdapter
-        binding.albumContentVp.orientation = ViewPager2.ORIENTATION_HORIZONTAL
-        TabLayoutMediator(binding.albumContentTb, binding.albumContentVp) {
-            tab, position ->
-            tab.text = tabInfo[position] // 탭뷰의 아이템 이름 설정
-        }.attach() // 탭뷰와 vp 연결
-
-        return binding.root
     }
 }
