+++ mission/week3/app/src/main/java/com/example/week3/ui/album/DetailFragment.kt
@@ -1,22 +1,22 @@
-package com.example.week3.fragment
+package com.example.week3.ui.album
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.week3.databinding.FragmentVideoBinding
+import com.example.week3.databinding.FragmentDetailBinding
 
-class VideoFragment : Fragment() {
+class DetailFragment : Fragment() {
 
-    lateinit var binding: FragmentVideoBinding
+    lateinit var binding: FragmentDetailBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentVideoBinding.inflate(layoutInflater)
+        binding = FragmentDetailBinding.inflate(layoutInflater)
 
         return binding.root
     }
