+++ mission/week3/app/src/main/java/com/example/week3/ui/album/SongFragment.kt
@@ -1,4 +1,4 @@
-package com.example.week3.fragment
+package com.example.week3.ui.album
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -24,6 +24,12 @@ class SongFragment : Fragment() {
             Toast.makeText(activity, "${song_title}이(가) 재생목록에 추가되었습니다.", Toast.LENGTH_SHORT).show()
         }
 
+        setupSongMixTgClickListener()
+
+        return binding.root
+    }
+
+    private fun setupSongMixTgClickListener() {
         binding.songMixonTg.setOnClickListener {
             setToggleSwitchStatus(true)
         }
@@ -31,11 +37,9 @@ class SongFragment : Fragment() {
         binding.songMixoffTg.setOnClickListener {
             setToggleSwitchStatus(false)
         }
-
-        return binding.root
     }
 
-    fun setToggleSwitchStatus(isToggleOn: Boolean) {
+    private fun setToggleSwitchStatus(isToggleOn: Boolean) {
         if (isToggleOn) {
             // 현재 값: 토글이 켜져있음
             // 기대 값: 토글 끄기
