+++ mission/week3/app/src/main/java/com/example/week3/ui/album/VideoFragment.kt
@@ -1,22 +1,22 @@
-package com.example.week3.fragment
+package com.example.week3.ui.album
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.week3.databinding.FragmentSearchBinding
+import com.example.week3.databinding.FragmentVideoBinding
 
-class SearchFragment : Fragment() {
+class VideoFragment : Fragment() {
 
-    lateinit var binding: FragmentSearchBinding
+    lateinit var binding: FragmentVideoBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View {
-        binding = FragmentSearchBinding.inflate(inflater, container, false)
+    ): View? {
+        binding = FragmentVideoBinding.inflate(layoutInflater)
 
         return binding.root
     }
