+++ mission/week3/app/src/main/java/com/example/week3/ui/home/BannerFragment.kt
@@ -1,22 +1,26 @@
-package com.example.week3.fragment
+package com.example.week3.ui.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.week3.databinding.FragmentLookBinding
+import com.example.week3.databinding.FragmentBannerBinding
 
-class LookFragment : Fragment() {
+class BannerFragment(
+    val imgRes : Int
+): Fragment() {
 
-    lateinit var binding: FragmentLookBinding
+    lateinit var binding: FragmentBannerBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentLookBinding.inflate(inflater, container, false)
+        binding = FragmentBannerBinding.inflate(layoutInflater)
+
+        binding.bannerImgIv.setImageResource(imgRes)
 
         return binding.root
     }
