+++ mission/week3/app/src/main/java/com/example/week3/ui/locker/DeviceMusicFragment.kt
@@ -1,22 +1,22 @@
-package com.example.week3.fragment
+package com.example.week3.ui.locker
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.week3.databinding.FragmentDetailBinding
+import com.example.week3.databinding.FragmentDeviceMusicBinding
 
-class DetailFragment : Fragment() {
+class DeviceMusicFragment: Fragment() {
 
-    lateinit var binding: FragmentDetailBinding
+    lateinit var binding: FragmentDeviceMusicBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentDetailBinding.inflate(layoutInflater)
+        binding = FragmentDeviceMusicBinding.inflate(layoutInflater)
 
         return binding.root
     }
