+++ mission/week3/app/src/main/java/com/example/week3/ui/locker/LockerVPAdapter.kt
@@ -1,9 +1,7 @@
-package com.example.week3.adapter
+package com.example.week3.ui.locker
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.example.week3.fragment.SavedMusicFragment
-import com.example.week3.fragment.DeviceMusicFragment
 
 class LockerVPAdapter(fragment: Fragment): FragmentStateAdapter(fragment) {
     override fun getItemCount(): Int = 2
