+++ mission/week3/app/src/main/java/com/example/week3/ui/main/LockerFragment.kt
@@ -1,4 +1,4 @@
-package com.example.week3.fragment
+package com.example.week3.ui.main
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.viewpager2.widget.ViewPager2
-import com.example.week3.adapter.LockerVPAdapter
+import com.example.week3.ui.locker.LockerVPAdapter
 import com.example.week3.databinding.FragmentLockerBinding
 import com.google.android.material.tabs.TabLayoutMediator
 
@@ -23,8 +23,11 @@ class LockerFragment : Fragment() {
         binding = FragmentLockerBinding.inflate(inflater, container, false)
 
         val lockerAdapter = LockerVPAdapter(this)
-        binding.lockerContentVp.adapter = lockerAdapter
-        binding.lockerContentVp.orientation = ViewPager2.ORIENTATION_HORIZONTAL
+        binding.lockerContentVp.apply {
+            adapter = lockerAdapter
+            orientation = ViewPager2.ORIENTATION_HORIZONTAL
+        }
+
         TabLayoutMediator(binding.lockerContentTb, binding.lockerContentVp) {
                 tab, position ->
             tab.text = tabInfo[position] // 탭뷰의 아이템 이름 설정
