+++ mission/week3/app/src/main/res/drawable/btm_color_selector.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:color="#3f3fff" android:state_checked="true" />
+    <item android:color="@color/black" android:state_checked="true" />
     <item android:color="#a8a8a8" />
 </selector>
