+++ mission/week3/app/src/main/res/layout/fragment_album.xml
@@ -22,7 +22,6 @@
         android:layout_height="35dp"
         android:layout_marginTop="40dp"
         android:layout_marginEnd="15dp"
-        android:background="@color/white"
         android:scaleType="fitXY"
         android:src="@drawable/btn_player_more"
         app:layout_constraintEnd_toEndOf="parent"
@@ -44,7 +43,6 @@
         android:layout_height="40dp"
         android:layout_marginStart="15dp"
         android:layout_marginTop="40dp"
-        android:background="@color/white"
         android:scaleType="fitXY"
         android:src="@drawable/btn_arrow_black"
         app:layout_constraintStart_toStartOf="parent"
@@ -112,6 +110,7 @@
         android:id="@+id/album_content_tb"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:background="@color/transparent"
         app:tabRippleColor="@color/transparent"
         app:tabIndicatorColor="@color/flo"
         app:tabSelectedTextColor="@color/flo"
