+++ mission/week3/app/src/main/res/layout/fragment_banner.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragment.BannerFragment">
+    tools:context=".ui.home.BannerFragment">
 
     <ImageView
         android:id="@+id/banner_img_iv"
