+++ mission/week3/app/src/main/res/layout/fragment_device_music.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragment.DeviceMusicFragment">
+    tools:context=".ui.locker.DeviceMusicFragment">
 
     <TextView
         android:layout_width="wrap_content"
