+++ mission/week3/app/src/main/res/layout/fragment_song.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragment.SongFragment">
+    tools:context=".ui.album.SongFragment">
 
 
     <androidx.core.widget.NestedScrollView
