+++ mission/week3/app/src/main/res/values/colors.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="gray">#D1D1D1</color>
+    <color name="light_gray">#F0F0F0</color>
 
     <color name="flo">#3f3fff</color>
